Internal_ffdf {ff}	R Documentation

Internal ffdf functions

Description

Internal ffdf functions

Usage

get_nvw(x)

Arguments

x
an object

Details

'get_nvw' gives the length and vw of an object (or its first dimension, if it is not a vector)

[Package ff version 4.5.2 Index]