array2vector {ff}	R Documentation

Array: make vector from array

Description

Makes a vector from an array respecting dim and dimorder

Usage

array2vector(x, dim = NULL, dimorder = NULL)

Arguments

x
an array

dim
dim

dimorder
dimorder

Details

This is the inverse function of vector2array. It extracts the vector from the array by first moving through the fastest rotating dimension dim[dimorder[1]], then dim[dimorder[2]], and so forth

Value

a vector

Author(s)

Jens Oehlschlägel

See Also

vector2array, arrayIndex2vectorIndex

Examples

  array2vector(matrix(1:12, 3, 4))
  array2vector(matrix(1:12, 3, 4, byrow=TRUE), dimorder=2:1)

[Package ff version 4.5.2 Index]