as.logical.bit(bit)	R Documentation

Coercion from bit

Description

Coercing from bit to logical, integer, which.

Usage


## S3 method for class 'bit':
as.logical(x, ...)
## S3 method for class 'bit':
as.integer(x, ...)
## S3 method for class 'bit':
as.double(x, ...)
## S3 method for class 'bit':
as.which(x, range = NULL, ...)

Arguments

x 	 an object of class bit

range 	 integer vector of length==2 giving a range restriction for chunked processing

... 	 further arguments (passed to which for the default method, ignored otherwise)

Details

as.logical.bit and as.integer.bit return a vector of FALSE, TRUE resp. 0,1. as.which.bit returns a vector of subscripts with class 'which' Coercion from bit is quite fast because we use a double loop that fixes each word in a processor register.

Value

a vector of class 'logical' or 'integer'

Author(s)

Jens Oehlschlägel

See Also

bit, as.bit, as.logical, as.integer, as.which, as.bitwhich, as.ff, as.hi

Examples


  x <- as.bit(c(FALSE, NA, TRUE, rep(TRUE, 9)))
  as.logical(x)
  as.integer(x)
  as.which.bit(x)

[Package bit version 1.1.0 Index]