bbatch(bit)	R Documentation

Balanced Batch sizes

Description

bbatch calculates batch sizes so that they have rather balanced sizes than very different sizes

Usage


bbatch(N, B)

Arguments

N 	 total size

B 	 desired batch size

Details

Tries to have rb==0 or rb as close to b as possible while guaranteing that rb < b && (b - rb) <= min(nb, b)

Value

a list with components

b 	 the batch size

nb 	 the number of batches

rb 	 the size of the rest

Author(s)

Jens Oehlschlägel

See Also

repfromto, ffvecapply

Examples


  bbatch(100, 24)

[Package bit version 1.1.0 Index]