chunk(bit)	R Documentation

Chunked range index

Description

creates a sequence of range indexes using a syntax not completely unlike 'seq'

Usage


chunk(from = NULL, to = NULL, by = NULL, length.out = NULL, along.with = NULL, overlap = 0L, method = c("bbatch", "seq"))

Arguments

from 	 the starting value of the sequence.

to 	 the (maximal) end value of the sequence.

by 	 increment of the sequence

length.out 	 desired length of the sequence.

along.with 	 take the length from the length of this argument.

overlap 	 number of values to overlap (will lower the starting value of the sequence, the first range becomes smaller

method 	 default 'bbatch' will try to balance the chunk size, see bbatch, 'seq' will create chunks like seq

Value

returns a list of ri objects representing chunks of subscripts

Author(s)

Jens Oehlschlägel

See Also

ri, seq, bbatch

Examples


  chunk(1, 100, by=30)
  chunk(1, 100, by=30, method="seq")

[Package bit version 1.1.0 Index]