close.ff {ff}	R Documentation

Closing ff files

Description

Close frees the Memory Mapping resources and closes the ff file without deleting the file data.

Usage

## S3 method for class 'ff'
close(con, ...)
## S3 method for class 'ffdf'
close(con, ...)
## S3 method for class 'ff_pointer'
close(con, ...)

Arguments

con
an open ff object

...
...

Details

The ff_pointer method is not intended for manual use, it is used at finalizer dispatch time. Closing ffdf objects will close all of their physical components including their row.names if they are is.ff

Value

TRUE if the file could be closed, FALSE if it was closed already (or NA if not all components of an ffdf returned FALSE or TRUE on closing)

Author(s)

Jens Oehlschlägel

See Also

ff, open.ff, delete, deleteIfOpen

Examples

  x <- ff(1:12)
  close(x)
  x
  open(x)
  x
  rm(x); gc()

[Package ff version 4.5.2 Index]