dimorderCompatible {ff}	R Documentation

Test for dimorder compatibility

Description

dimorderStandard returns TRUE if the dimorder is standard (ascending), vectorStandard returns TRUE if the dimorder-bydim combination is compatible with a standard elementwise vector interpretation, dimorderCompatible returns TRUE if two dimorders have a compatible elementwise vector interpretation and vectorCompatible returns TRUE if dimorder-bydim combinations have a compatible elementwise vector interpretation.

Usage

dimorderStandard(dimorder)
vectorStandard(dimorder, bydim = NULL)
dimorderCompatible(dim, dim2, dimorder, dimorder2)
vectorCompatible(dim, dim2, dimorder=NULL, dimorder2=NULL, bydim = NULL, bydim2 = NULL)

Arguments

dim
a dim

dim2
a dim

dimorder
a dimorder

dimorder2
a dimorder

bydim
a bydim order, see [.ff

bydim2
a bydim order, see argument fromdim in update.ff

Value

TRUE if compatibility has been detected, FALSE otherwise

Note

does not yet gurantee to detect all compatible configurations, but the most important ones

Author(s)

Jens Oehlschlägel

See Also

dimorder, ffconform

[Package ff version 4.5.2 Index]