ffinfo {ff}	R Documentation

Inspect content of ff saves

Description

Find out which objects and ff files are in a pair of files saved with ffsave

Usage

ffinfo(file)

Arguments

file
a character string giving the name (without extension) of the .RData and .ffData files to load

Value

a list with components

RData
a list, one element for each object (named like the object): a character vector with the names of the ff files

ffData
a list, one element for each path (names like the path): a character vector with the names of the ff files

rootpath
the root path relative to which the files are stored in the .ffData zip

Note

For large files and the zip64 format use zip 3.0 and unzip 6.0 from https://infozip.sourceforge.net/.

Author(s)

Jens Oehlschlägel

See Also

ffsave, ffload, ffdrop

[Package ff version 4.5.2 Index]