ffload {ff}	R Documentation

Reload ffSaved Datasets

Description

Reload datasets written with the function ffsave or ffsave.image .

Usage

ffload(file, list = character(0L), envir = parent.frame()
, rootpath = NULL, overwrite = FALSE)

Arguments

file
a character string giving the name (without extension) of the .RData and .ffData files to load

list
An optional vector of names selecting those objects to be restored (default NULL restores all)

envir
the environment where the data should be loaded.

rootpath
an optional rootpath where to restore the ff files (default NULL restores in the original location)

overwrite
logical indicating whether possibly existing ff files shall be overwritten

Details

ffinfo can be used to inspect the contents an ffsaved pair of .RData and .ffData files. Argument list can then be used to restore only part of the ffsave.

Value

A character vector with the names of the restored ff files

Note

The ff files are not platform-independent with regard to byte order. For large files and the zip64 format use zip 3.0 and unzip 6.0 from https://infozip.sourceforge.net//.

Author(s)

Jens Oehlschlägel

See Also

load, ffsave, ffinfo, ffdrop

[Package ff version 4.5.2 Index]