ffreturn {ff}	R Documentation

Return suitable ff object

Description

ffreturn returns FF_RETURN if it is ffsuitable otherwise creates a suitable ffsuitable object

Usage

ffreturn(FF_RETURN = NULL, FF_PROTO = NULL, FF_ATTR = NULL)

Arguments

FF_RETURN
the object to be tested for suitability

FF_PROTO
the prototype object which FF_RETURN should match

FF_ATTR
a list of additional attributes dominating those from FF_PROTO

Value

a suitable ffsuitable object

Note

xx Work in progress for package R.ff

Author(s)

Jens Oehlschlägel

See Also

ffconform, ffsuitable

[Package ff version 4.5.2 Index]