ffsuitable {ff}	R Documentation

Test ff object for suitability

Description

ffsuitable tests whether FF_RETURN is an ff object like FF_PROTO and having attributes FF_ATTR.

Usage

ffsuitable(FF_RETURN, FF_PROTO = NULL, FF_ATTR = list()
, strict.dimorder = TRUE, fail = "warning")
ffsuitable_attribs(x)

Arguments

x
an object from which to extract attributes for comparison

FF_RETURN
the object to be tested for suitability

FF_PROTO
the prototype object which FF_RETURN should match

FF_ATTR
a list of additional attributes dominating those from FF_PROTO

strict.dimorder
if TRUE ffsuitability requires that the dimorders are standard (ascending)

fail
name of a function to be called if not ffsuitable (default warning)

Value

TRUE if FF_RETURN object is suitable, FALSE otherwise

Note

xx Work in progress for package R.ff

Author(s)

Jens Oehlschlägel

See Also

ffconform, ffreturn

[Package ff version 4.5.2 Index]