ffxtensions {ff}	R Documentation

Test for availability of ff extensions

Description

checks if this version of package ff supports ff extensions.

Usage

 ffxtensions()
 ffsymmxtensions()

Details

ff extensions are needed for certain bitcompressed vmodes and ff symm extensions for symmetric matrices.

Value

logical scalar

Author(s)

Jens Oehlschlägel

See Also

vmode

Examples

  ffxtensions()
  ffsymmxtensions()

[Package ff version 4.5.2 Index]