getpagesize {ff}	R Documentation

Get page size information

Description

The function is used for obtaining the natural OS-specific page size in Bytes. getpagesize returns the OS-specific page size in Bytes for memory mapped files, while getdefaultpagesize returns a suggested page size. getalignedpagesize returns the pagesize as a multiple of the OS-specific page size in Bytes, which is the correct way to specify pagesize in ff.

Usage

  getpagesize()
  getdefaultpagesize()
  getalignedpagesize(pagesize)
 

Arguments

pagesize
a desired pagesize in bytes

Value

An integer giving the page size in Bytes.

Author(s)

Daniel Adler, Jens Oehlschlägel

Examples

  getpagesize()
  getdefaultpagesize()
  getalignedpagesize(2000000)
  

[Package ff version 4.5.2 Index]