is.open {ff}	R Documentation

Test if object is opened

Description

Test whether an ff or ffdf object or a ff_pointer is opened.

Usage

is.open(x, ...)
## S3 method for class 'ff'
is.open(x, ...)
## S3 method for class 'ffdf'
is.open(x, ...)
## S3 method for class 'ff_pointer'
is.open(x, ...)

Arguments

x
an ff or ffdf object

...
further arguments (not used)

Details

ff objects open automatically if accessed while closed. For ffdf objects we test all of their physical components including their row.names if they are is.ff

Value

TRUE or FALSE (or NA if not all components of an ffdf object are opened or closed)

Author(s)

Jens Oehlschlägel

See Also

is.readonly, open.ff, close.ff

Examples

  x <- ff(1:12)
  is.open(x)
  close(x)
  is.open(x)
  rm(x); gc()

[Package ff version 4.5.2 Index]