is.readonly {ff}	R Documentation

Get readonly status

Description

Get readonly status of an ff object

Usage

is.readonly(x, ...)
## S3 method for class 'ff'
is.readonly(x, ...)

Arguments

x
x

...
...

Details

ff objects can be created/opened with readonly=TRUE. After each opening of the ff file readonly status is stored in the physical attributes and serves as the default for the next opening. Thus querying a closed ff object gives the last readonly status.

Value

logical scalar

Author(s)

Jens Oehlschlägel

See Also

open.ff, physical

Examples

  x <- ff(1:12)
  is.readonly(x)
  close(x)
  open(x, readonly=TRUE)
  is.readonly(x)
  close(x)
  is.readonly(x)
  rm(x)

[Package ff version 4.5.2 Index]