maxffmode {ff}	R Documentation

Lossless vmode coercability

Description

maxffmode returns the lowest vmode that can absorb all input vmodes without data loss

Usage

maxffmode(...)

Arguments

...
one or more vectors of vmodes

Value

the smallest .ffmode which can absorb the input vmodes without data loss

Note

The output can be larger than any of the inputs (if the highest input vmode is an integer type without NA and any other input requires NA).

Author(s)

Jens Oehlschlägel

See Also

.vcoerceable, .ffmode, ffconform

Examples

   maxffmode(c("quad","logical"), "ushort")

[Package ff version 4.5.2 Index]