maxlength {ff}	R Documentation

Get physical length of an ff or ram object

Description

maxlength returns the physical length of an ff or ram object

Usage

maxlength(x, ...)
## S3 method for class 'ff'
maxlength(x, ...)
## Default S3 method:
maxlength(x, ...)

Arguments

x
ff or ram object

...
additional arguments (not used)

Value

integer scalar

Author(s)

Jens Oehlschlägel

See Also

length.ff, maxindex

Examples

  x <- ff(1:12)
  length(x) <- 10
  length(x)
  maxlength(x)
  x
  rm(x); gc()

[Package ff version 4.5.2 Index]