mismatch {ff}	R Documentation

Test for recycle mismatch

Description

mismatch will return TRUE if the larger of nx,ny is not a multiple of the other and the other is >0 (see arithmetic.c). ymismatch will return TRUE if nx is not a multiple of ny and ny>0

Usage

mismatch(nx, ny)
ymismatch(nx, ny)

Arguments

nx
x length

ny
y length

Value

logical scalar

Author(s)

Jens Oehlschlägel

See Also

ffconform

Examples

  ymismatch(4,0)
  ymismatch(4,2)
  ymismatch(4,3)
  ymismatch(2,4)
  mismatch(4,0)
  mismatch(4,2)
  mismatch(4,3)
  mismatch(2,4)

[Package ff version 4.5.2 Index]