names.ff {ff}	R Documentation

Getting and setting names

Description

For ff_vectors you can set names, though this is not recommended for large objects.

Usage

  ## S3 method for class 'ff'
names(x)
  ## S3 replacement method for class 'ff'
names(x) <- value
  ## S3 method for class 'ff_array'
names(x)
  ## S3 replacement method for class 'ff_array'
names(x) <- value

Arguments

x
a ff vector

value
a character vector

Details

If vw is set, names.ff returns the appropriate part of the names, but you can't set names while vw is set. names.ff_array returns NULL and setting names for ff_arrays is not allowed, but setting dimnames is.

Value

names returns a character vector (or NULL)

Author(s)

Jens Oehlschlägel

See Also

names, dimnames.ff_array, vw, virtual

Examples

  x <- ff(1:26, names=letters)
  names(x)
  names(x) <- LETTERS
  names(x)
  names(x) <- NULL
  names(x)
  rm(x); gc()

[Package ff version 4.5.2 Index]