open.ff {ff}	R Documentation

Opening an ff file

Description

open.ff opens an ff file, optionally marking it readonly and optionally specifying a caching scheme.

Usage

 ## S3 method for class 'ff'
open(con, readonly = FALSE, pagesize = NULL, caching = NULL, assert = FALSE, ...)
 ## S3 method for class 'ffdf'
open(con, readonly = FALSE, pagesize = NULL, caching = NULL, assert = FALSE, ...)

Arguments

con
an ff or ffdf object

readonly
readonly

pagesize
number of bytes to use as pagesize or NULL to take the pagesize stored in the physical attribute of the ff object, see getalignedpagesize

caching
one of 'mmnoflush' or 'mmeachflush', see ff

assert
setting this to TRUE will give a message if the ff was not open already

...
further arguments (not used)

Details

ff objects will be opened automatically when accessing their content and the file is still closed. Opening ffdf objects will open all of their physical components including their row.names if they are is.ff

Value

TRUE if object could be opened, FALSE if it was opened already (or NA if not all components of an ffdf returned FALSE or TRUE on opening)

Author(s)

Jens Oehlschlägel

See Also

ff, close.ff, delete, deleteIfOpen, getalignedpagesize

Examples

  x <- ff(1:12)
  close(x)
  is.open(x)
  open(x)
  is.open(x)
  close(x)
  is.open(x)
  x[]
  is.open(x)
  y <- x
  close(y)
  is.open(x)
  rm(x,y); gc()

[Package ff version 4.5.2 Index]