physical(bit)	R Documentation

Physical and virtual attributes

Description

Compatibility functions (to package ff) for getting and setting physical and virtual attributes.

Usage


physical(x)
virtual(x)
physical(x) <- value
virtual(x) <- value
## Default S3 method:
physical(x)
## Default S3 method:
virtual(x)
## Default S3 method:
physical(x) <- value
## Default S3 method:
virtual(x) <- value
## S3 method for class 'physical':
print(x, ...)
## S3 method for class 'virtual':
print(x, ...)

Arguments

x 	 a ff or ram object

value 	 a list with named elements

... 	 further arguments

Details

ff objects have physical and virtual attributes, which have different copying semantics: physical attributes are shared between copies of ff objects while virtual attributes might differ between copies. as.ram will retain some physical and virtual atrributes in the ram clone, such that as.ff can restore an ff object with the same attributes.

Value

physical and virtual returns a list with named elements

Author(s)

Jens Oehlschlägel

See Also

physical

Examples


  physical(bit(12))
  virtual(bit(12))

[Package bit version 1.1.0 Index]