physical.ff {ff}	R Documentation

Getting and setting physical and virtual attributes of ff objects

Description

Functions for getting and setting physical and virtual attributes of ff objects.

Usage

## S3 method for class 'ff'
physical(x)
## S3 method for class 'ff'
virtual(x)
## S3 replacement method for class 'ff'
physical(x) <- value
## S3 replacement method for class 'ff'
virtual(x) <- value

Arguments

x
an ff object

value
a list with named elements

Details

ff objects have physical and virtual attributes, which have different copying semantics: physical attributes are shared between copies of ff objects while virtual attributes might differ between copies. as.ram will retain some physical and virtual atrributes in the ram clone, such that as.ff can restore an ff object with the same attributes.

Value

physical and virtual returns a list with named elements

Author(s)

Jens Oehlschlägel

See Also

physical.ff, physical.ffdf, ff, as.ram;
is.sorted and na.count for applications of physical attributes;
levels.ff and ramattribs for applications of virtual attributes

Examples

  x <- ff(1:12)
  x
  physical(x)
  virtual(x)
  y <- as.ram(x)
  physical(y)
  virtual(y)
  rm(x,y); gc()

[Package ff version 4.5.2 Index]