print.ff {ff}	R Documentation

Print and str methods

Description

printing ff objects and compactly showing their structure

Usage

## S3 method for class 'ff'
print(x, ...)
## S3 method for class 'ff_vector'
print(x, maxlength = 16, ...)
## S3 method for class 'ff_matrix'
print(x, maxdim = c(16, 16), ...)
## S3 method for class 'ff'
str(object, nest.lev=0, ...)
## S3 method for class 'ffdf'
str(object, nest.lev=0, ...)

Arguments

x
a ff object

object
a ff object

nest.lev
current nesting level in the recursive calls to str

maxlength
max number of elements to print from an ff_vector

maxdim
max number of elements to print from each dimension from an ff_array

...
further arguments to print

Details

The print methods just print a few exmplary elements from the beginning and end of the dimensions.

Value

invisible()

Author(s)

Jens Oehlschlägel

See Also

ff, print, str

Examples

  x <- ff(1:10000)
  x
  print(x, maxlength=30)
  dim(x) <- c(100,100)
  x
  rm(x); gc()

[Package ff version 4.5.2 Index]