ramattribs {ff}	R Documentation

Get ramclass and ramattribs

Description

Functions ramclass and ramattribs return the respective virtual attributes, that determine which class (and attributes) an ff object receives when subscripted (or coerced) to ram.

Usage

ramclass(x, ...)
## S3 method for class 'ff'
ramclass(x, ...)
## Default S3 method:
ramclass(x, ...)
ramattribs(x, ...)
## S3 method for class 'ff'
ramattribs(x, ...)
## Default S3 method:
ramattribs(x, ...)

Arguments

x
x

...
further arguments (not used)

Details

ramclass and ramattribs provide a general mechanism to store atomic classes in ff objects, for example factor – see levels.ff – and POSIXct, see the example.

Value

ramclass returns a character vector with classnames and ramattribs returns a list with names elemens just like attributes. The vectors ramclass_excludes and ramattribs_excludes name those attributes, which are not exported from ff to ram objects when using as.ram.

Author(s)

Jens Oehlschlägel

See Also

ff, virtual, as.ram, levels.ff, attributes, DateTimeClasses

Examples

  x <- ff(as.POSIXct(as.POSIXlt(Sys.time(), "GMT")), length=12)
  x
  ramclass(x)
  ramattribs(x)
  class(x[])
  attributes(x[])
  virtual(x)$ramattribs$tzone = NULL
  attributes(x[])
  rm(x); gc()

[Package ff version 4.5.2 Index]