ri(bit)	R Documentation

Range index

Description

A range index can be used to extract or replace a continuous ascending part of the data

Usage


ri(from, to = NULL, maxindex=NA)
## S3 method for class 'ri':
print(x, ...)


Arguments

from 	 first position

to 	 last posistion

x 	 an object of class 'ri'

maxindex 	 the maximal length of the object-to-be-subscripted (if known)

... 	 further arguments

Value

A two element integer vector with class 'ri'

Author(s)

Jens Oehlschlägel

See Also

as.hi.ri

Examples


 bit(12)[ri(1,6)]

[Package bit version 1.1.0 Index]