unsort {ff}	R Documentation

Hybrid Index, internal utilities

Description

Non-documented internal utilities that might change

Usage

unsort(x, ix)
unsort.hi(x, index)
unsort.ahi(x, index, ixre = any(sapply(index, function(i) {
    if (is.null(i$ix)) {
        if (i$re) TRUE else FALSE
    } else {
        TRUE
    }
})), ix = lapply(index, function(i) {
    if (is.null(i$ix)) {
        if (i$re)
            orig <- rev(seq_len(poslength(i)))
        else orig <- seq_len(poslength(i))
    }
    else {
        orig <- i$ix
    }
    orig
}))
subscript2integer(x, maxindex = NULL, names = NULL)

Arguments

x
x

ix
ix

ixre
ixre

index
index

maxindex
maxindex

names
names

Details

These are utility functions for restoring original order after sorting. For now we 'mimic' the intuitive but wrong argument order of match() which should rather have the 'table' argument as its first argument, then one could properly method-dispatch on the type of table. xx We might change to proper 'unsort' generic, but then we have to change argument order.

Value

undefined

Author(s)

Jens Oehlschlägel

See Also

hi, as.hi

[Package ff version 4.5.2 Index]