vmode.ffdf {ff}	R Documentation

Virtual storage mode of ffdf

Description

Function vmode returns the virtual storage mode of each ffdf column

Usage

## S3 method for class 'ffdf'
vmode(x, ...)

Arguments

x
ffdf

...
ignored

Value

a character vector with one element for each column

Author(s)

Jens Oehlschlägel

See Also

vmode, ffdf

Examples

  vmode(as.ffdf(data.frame(a=as.double(1:26), b=letters, stringsAsFactors = TRUE)))
  gc()

[Package ff version 4.5.2 Index]