vt {ff}	R Documentation

Virtual transpose

Description

The vt generic does a matrix or array transpose by modifying virtual attributes rather than by physically copying matrix elements.

Usage

vt(x, ...)
## S3 method for class 'ff'
vt(x, ...)
## Default S3 method:
vt(x, ...)
## S3 method for class 'ff'
t(x)

Arguments

x
an ff or ram object

...
further arguments (not used)

Details

The vt.ff method does transpose through reversing dim.ff and dimorder. The vt.default method is a wrapper to the standard transpose t.
The t.ff method creates a transposed clone.
If x has a virtual window vw defined, vt.ff returns an ff object with a transposed virtual window, the t.ff method return a transposed clone of the virtual window content only.

Value

an object that behaves like a transposed matrix

Author(s)

Jens Oehlschlägel

See Also

dim.ff, vw, virtual

Examples

  x <- ff(1:20, dim=c(4,5))
  x
  vt(x)
  y <- t(x)
  y
  vw(x) <- cbind(c(1,3,0),c(1,4,0))
  x
  vt(x)
  y <- t(x)
  y
  rm(x,y); gc()

[Package ff version 4.5.2 Index]